scripts/azure-block-unused-resources/requirements.txt
azure-common
azure-identity
azure-mgmt-core
azure-mgmt-resource
azure-mgmt-resourcegraph
azure-mgmt-subscription
azure-mgmt-policyinsights
PyGithub
requests
psutil